backend/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-filter==23.1
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.3
psycopg2==2.9.5
pytz==2023.2
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
